refactor: Update attribute casting and method visibility across multiple models
- Refactored attribute casting to use the new #[\Override] attribute for consistency.
- Changed method visibility from public to protected for various methods to encapsulate functionality.
- Improved type hinting and return types for better clarity and type safety across models.

+++ app/Models/AfipImportacionCrudaModel.php
@@ -73,7 +73,7 @@ class AfipImportacionCrudaModel extends Model
             }
 
             // Insertar las líneas restantes
-            if (!empty($batchData)) {
+            if ($batchData !== []) {
                 DB::connection($this->connection)
                     ->table($this->table)
                     ->insert($batchData);
@@ -118,12 +118,10 @@ class AfipImportacionCrudaModel extends Model
     private function truncateTableIfNotEmpty(): void
     {
         // Verificar si la tabla existe
-        if (DB::connection($this->connection)->getSchemaBuilder()->hasTable($this->table)) {
-            // Verificar si la tabla no esta vacia
-            if (DB::connection($this->connection)->table($this->table)->count() > 0) {
-                // Eliminar la tabla
-                DB::connection($this->connection)->table($this->table)->truncate();
-            }
+        // Verificar si la tabla no esta vacia
+        if (DB::connection($this->connection)->getSchemaBuilder()->hasTable($this->table) && DB::connection($this->connection)->table($this->table)->count() > 0) {
+            // Eliminar la tabla
+            DB::connection($this->connection)->table($this->table)->truncate();
         }
     }
 }
